Shift components to their own file
