+++ TourMate/TourMate/Backend/Network/Models/APICall.swift
@@ -14,28 +14,6 @@ protocol APICall {
     func body() -> Data?
 }
 
-enum APIError: Swift.Error {
-    case invalidURL
-    case httpCode(HTTPCode)
-    case unexpectedResponse
-    case imageProcessing([URLRequest])
-}
-
-extension APIError: LocalizedError {
-    var errorDescription: String? {
-        switch self {
-        case .invalidURL:
-            return "Invalid URL"
-        case let .httpCode(code):
-            return "Unexpected HTTP code: \(code)"
-        case .unexpectedResponse:
-            return "Unexpected response from the server"
-        case .imageProcessing:
-            return "Unable to load image"
-        }
-    }
-}
-
 extension APICall {
     func urlRequest(baseURL: String) throws -> URLRequest {
         guard let url = URL(string: baseURL + path) else {
