+++ TourMate/TourMate/Backend/Network/Models/APIError.swift
@@ -1,0 +1,30 @@
+//
+//  APIError.swift
+//  TourMate
+//
+//  Created by Tan Rui Quan on 31/3/22.
+//
+
+import Foundation
+
+enum APIError: Swift.Error {
+    case invalidURL
+    case httpCode(HTTPCode)
+    case unexpectedResponse
+    case imageProcessing([URLRequest])
+}
+
+extension APIError: LocalizedError {
+    var errorDescription: String? {
+        switch self {
+        case .invalidURL:
+            return "Invalid URL"
+        case let .httpCode(code):
+            return "Unexpected HTTP code: \(code)"
+        case .unexpectedResponse:
+            return "Unexpected response from the server"
+        case .imageProcessing:
+            return "Unable to load image"
+        }
+    }
+}
